+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Update to Bevy 0.18.0-rc.1.
+- Change `TimeKind::default` to `TimeKind::Real`.
+- Add `TimeKind` configuration to `SmoothNudge`.
+- Add `TimeKind` configuration to `DeltaScale`.
+- `SmoothNudge::new` no longer `const`.
 
 ## [0.20.0] - 2025-11-15